Move MetricClient.Listener to MetricListener
